requirements.txt
tensorflow_gpu==2.2.0
imutils==0.5.4
opencv_python==4.5.3.56
scikit_image==0.18.3
scikit_learn==0.24.2
numpy==1.21.2
py_sudoku==1.0.1


